Remove the old way of create the preconfig and put the new way in place
